feat: setup schedule wakeup command
